stepfunctions-data-science-sdk/cfn2/requirements.txt
aws-cdk.core>=1.170.0
aws-cdk.assertions>=1.170.0
aws_cdk.aws_stepfunctions>=1.170.0
aws_cdk.aws_stepfunctions_tasks>=1.170.0
aws_cdk.aws_lambda>=1.170.0
aws_cdk.aws_iam>=1.170.0
aws_cdk.aws_s3>=1.170.0
aws_cdk.aws_ec2>=1.170.0
aws_cdk.aws_sqs>=1.170.0
aws_cdk.aws_lambda_event_sources>=1.170.0
aws_cdk.aws_glue>=1.170.0
boto3
sagemaker
